ud220915

+++ backend/controllers/articles.js
@@ -1,4 +1,5 @@
 const Article = require("../models/Article");
+const User = require("../models/User");
 const fs = require("fs");
 
 exports.createArticle = (req, res, next) => {
@@ -88,18 +89,18 @@ exports.modifyArticle = (req, res, next) => {
 exports.deleteArticle = (req, res, next) => {
   Article.findOne({ _id: req.params.id })
     .then((article) => {
-      // if (article.userId != req.auth.userId || !req.user.isAdmin) {
-      //   res.status(401).json({ message: "Not authorized" });
-      // } else {
-      const filename = article.imageUrl.split("/images/")[1];
-      fs.unlink(`images/${filename}`, () => {
-        Article.deleteOne({ _id: req.params.id })
-          .then(() => {
-            res.status(204).json({ message: "Deleted successfully!" });
-          })
-          .catch((error) => res.status(401).json({ error }));
-      });
-      // }
+      if (article.userId != req.auth.userId && !req.auth.isAdmin) {
+        res.status(401).json({ message: "Not authorized" });
+      } else {
+        const filename = article.imageUrl.split("/images/")[1];
+        fs.unlink(`images/${filename}`, () => {
+          Article.deleteOne({ _id: req.params.id })
+            .then(() => {
+              res.status(204).json({ message: "Deleted successfully!" });
+            })
+            .catch((error) => res.status(401).json({ error }));
+        });
+      }
     })
     .catch((error) => {
       res.status(500).json({ error });
